+++ .dockerignore
@@ -34,6 +34,7 @@ tests
 .env.test.local
 .env.production.local
 dist
+dist-verify*
 coverage
 build
