+++ src/App.vue
@@ -4,13 +4,5 @@
 </template>
 
 <script setup lang="ts">
-import { onMounted } from 'vue';
-import { useApiStore } from '@api/store.ts';
 import NavPartial from '@partials/NavPartial.vue';
-
-const apiStore = useApiStore();
-
-onMounted(() => {
-	apiStore.boot();
-});
 </script>
