+++ src/support/content/about-page.ts
@@ -1,4 +1,4 @@
 import aboutPage from '@fixtures/about-page.json';
-import type { AboutPageContent } from '@support/content.ts';
+import type { AboutPageContent } from '@support/content-types.ts';
 
 export const aboutPageContent: AboutPageContent = aboutPage;
