+++ src/support/content/contact-page.ts
@@ -1,4 +1,4 @@
 import contactPage from '@fixtures/contact-page.json';
-import type { ContactPageContent } from '@support/content.ts';
+import type { ContactPageContent } from '@support/content-types.ts';
 
 export const contactPageContent: ContactPageContent = contactPage;
