+++ src/support/content/home-page.ts
@@ -1,4 +1,4 @@
 import homePage from '@fixtures/home-page.json';
-import type { HomePageContent } from '@support/content.ts';
+import type { HomePageContent } from '@support/content-types.ts';
 
-export const homePageContent: HomePageContent = homePage;
+export const homePageContent = homePage as HomePageContent;
