+++ src/support/content/projects-page.ts
@@ -1,4 +1,4 @@
 import projectsPage from '@fixtures/projects-page.json';
-import type { ProjectsPageContent } from '@support/content.ts';
+import type { ProjectsPageContent } from '@support/content-types.ts';
 
 export const projectsPageContent: ProjectsPageContent = projectsPage;
