+++ src/support/content/terms-and-policies-page.ts
@@ -1,4 +1,4 @@
 import termsAndPoliciesPage from '@fixtures/terms-and-policies-page.json';
-import type { TermsAndPoliciesPageContent } from '@support/content.ts';
+import type { TermsAndPoliciesPageContent } from '@support/content-types.ts';
 
 export const termsAndPoliciesPageContent: TermsAndPoliciesPageContent = termsAndPoliciesPage;
