+++ src/support/content/work-with-us-page.ts
@@ -1,4 +1,4 @@
 import workWithUsPage from '@fixtures/work-with-us-page.json';
-import type { WorkWithUsPageContent } from '@support/content.ts';
+import type { WorkWithUsPageContent } from '@support/content-types.ts';
 
 export const workWithUsPageContent: WorkWithUsPageContent = workWithUsPage;
