+++ src/support/content/writing-page.ts
@@ -1,4 +1,4 @@
 import writingPage from '@fixtures/writing-page.json';
-import type { WritingPageContent } from '@support/content.ts';
+import type { WritingPageContent } from '@support/content-types.ts';
 
 export const writingPageContent: WritingPageContent = writingPage;
