+++ src/support/links.ts
@@ -14,20 +14,30 @@ interface LinksNavLink {
 }
 
 type PlatformName = 'x' | 'youtube' | 'instagram' | 'linkedin' | 'github';
-type NavSocialPlatform = Extract<PlatformName, 'linkedin' | 'x' | 'github'>;
+export type NavSocialPlatform = Extract<PlatformName, 'linkedin' | 'x' | 'github'>;
 
 export const NAV_SOCIAL_FALLBACKS: Record<NavSocialPlatform, string> = {
 	linkedin: 'https://www.linkedin.com/in/gocanto/',
 	x: 'https://x.com/oullin',
 	github: 'https://github.com/oullin',
 };
+export const NAV_SOCIAL_ORDER: NavSocialPlatform[] = ['linkedin', 'x', 'github'];
 
 const isHttpUrl = (url: string): boolean => /^https?:\/\//i.test(url);
+
+export const buildNavSocialLinkEntries = (links: Record<NavSocialPlatform, string>): LinksResponse[] =>
+	NAV_SOCIAL_ORDER.map((platform) => ({
+		uuid: `social-${platform}`,
+		name: platform,
+		handle: '',
+		url: links[platform],
+		description: '',
+	}));
 
 export const resolveNavSocialLinks = (links: LinksResponse[]): Record<NavSocialPlatform, string> => {
 	const resolvedLinks = { ...NAV_SOCIAL_FALLBACKS };
 
-	for (const platform of Object.keys(NAV_SOCIAL_FALLBACKS) as NavSocialPlatform[]) {
+	for (const platform of NAV_SOCIAL_ORDER) {
 		const match = links.find((item) => item.name === platform && item.url);
 		if (match?.url && isHttpUrl(match.url)) {
 			resolvedLinks[platform] = match.url;
